+++ tanaguru-engine/tanaguru-api/src/main/java/com/tanaguru/service/ProjectService.java
@@ -16,7 +16,7 @@ import java.util.Collection;
 import java.util.Optional;
 
 public interface ProjectService {
-    Project createProject(Contract contract, String name, String domain, boolean allowPageAudit, boolean allowSiteAudit, boolean allowScenarioAudit, boolean allowUploadAudit);
+    Project createProject(Contract contract, String name, String domain, boolean allowPageAudit, boolean allowSiteAudit, boolean allowScenarioAudit, boolean allowUploadAudit, Boolean isTrial);
 
     /**
      * Find authorities for a given @see EProjectRole